REFACTOR: move native parameter-binding cache into param_bind_cache.hpp
The handle-owned parameter-binding reuse cache was added inline to the ~6,000
line ddbc_bindings.cpp monolith and its struct definitions to ddbc_bindings.h.
It is a self-contained unit, so move it into its own single-responsibility
header, matching how the perf work already carved parameter detection into
param_detect.hpp, the type cache into py_type_cache.hpp, and the refcount
helpers into py_ref.hpp.

param_bind_cache.hpp now owns ParameterBinding, ExecuteBindingCache,
ExecuteParamBuffers, the single-buffer AllocateParamBuffer template and its
in-place-reuse overload, UpdateParamBuffer, SameParameterShape, and
CanCacheParameters. ddbc_bindings.h keeps only a forward declaration of
ExecuteBindingCache, which is all SqlHandle needs since it holds a unique_ptr to
it and defines every method that touches it out of line; this also avoids
pulling param_detect.hpp (which includes ddbc_bindings.h back) into the handle
header. AllocateParamBufferArray stays in ddbc_bindings.cpp because it serves the
executemany array-binding path, not the reuse cache.

Pure relocation with no behavior change: the moved code is byte-identical, the
release universal2 build is clean, and the full suite is unchanged (2344 passed,
the lone test_004 stored-procedure failure is a pre-existing stale-DB-object
artifact that passes in isolation).
