+++ mssql_python/pybind/ddbc_bindings.cpp
@@ -9,6 +9,7 @@
 #include "connection/connection_pool.h"
 #include "logger_bridge.hpp"
 #include "param_detect.hpp"
+#include "param_bind_cache.hpp"
 #include "py_ref.hpp"
 #include "py_type_cache.hpp"
 #include "utf_utils.h"
@@ -305,97 +306,11 @@ std::string MakeParamMismatchErrorStr(const SQLSMALLINT cType, const int paramIn
     return errorString;
 }
 
-// This function allocates a buffer of ParamType, stores it as a void* in
-// paramBuffers for book-keeping and then returns a ParamType* to the allocated
-// memory. ctorArgs are the arguments to ParamType's constructor used while
-// creating/allocating ParamType
-template <typename ParamType, typename... CtorArgs>
-ParamType* AllocateParamBuffer(std::vector<std::shared_ptr<void>>& paramBuffers,
-                               CtorArgs&&... ctorArgs) {
-    paramBuffers.emplace_back(new ParamType(std::forward<CtorArgs>(ctorArgs)...),
-                              std::default_delete<ParamType>());
-    LOG("AllocateParamBuffer: New owned buffer");
-    return static_cast<ParamType*>(paramBuffers.back().get());
-}
-
-struct ExecuteParamBuffers {
-    std::vector<std::shared_ptr<void>>& current;
-    const std::vector<std::shared_ptr<void>>* previous;
-};
-
-template <typename T>
-static bool UpdateParamBuffer(T& target, T&& value) {
-    target = std::move(value);
-    return true;
-}
-
-template <typename Char>
-static bool UpdateParamBuffer(std::basic_string<Char>& target, std::basic_string<Char>&& value) {
-    // Equal byte lengths keep both the address and ODBC BufferLength unchanged.
-    // Do not assign a string: even a capacity-preserving assignment may move it.
-    if (target.size() != value.size())
-        return false;
-    std::copy(value.begin(), value.end(), target.begin());
-    return true;
-}
-
-template <typename ParamType, typename... CtorArgs>
-ParamType* AllocateParamBuffer(ExecuteParamBuffers& buffers, CtorArgs&&... ctorArgs) {
-    ParamType value(std::forward<CtorArgs>(ctorArgs)...);
-    const size_t index = buffers.current.size();
-    if (buffers.previous && index < buffers.previous->size()) {
-        const auto& previous = (*buffers.previous)[index];
-        auto* target = static_cast<ParamType*>(previous.get());
-        if (UpdateParamBuffer(*target, std::move(value))) {
-            buffers.current.push_back(previous);
-            return target;
-        }
-    }
-    return AllocateParamBuffer<ParamType>(buffers.current, std::move(value));
-}
-
-static bool SameParameterShape(const ParameterBinding& binding, const ParamInfo& info) {
-    return binding.direction == info.inputOutputType && binding.cType == info.paramCType &&
-           binding.sqlType == info.paramSQLType && binding.columnSize == info.columnSize &&
-           binding.scale == info.decimalDigits;
-}
-
-static bool CanCacheParameters(const std::vector<ParamInfo>& infos) {
-    // Bound retained storage to SQL Server's scalar parameter limit. NULL/DAE and
-    // descriptor-based/complex types deliberately use the existing uncached path.
-    if (infos.empty() || infos.size() > 2100)
-        return false;
-    for (const auto& info : infos) {
-        if (info.isDAE || info.inputOutputType != SQL_PARAM_INPUT)
-            return false;
-        switch (info.paramCType) {
-            case SQL_C_CHAR:
-            case SQL_C_WCHAR:
-            case SQL_C_BINARY:
-                if (info.columnSize > MAX_INLINE_BINARY)
-                    return false;
-                break;
-            case SQL_C_BIT:
-            case SQL_C_STINYINT:
-            case SQL_C_TINYINT:
-            case SQL_C_SSHORT:
-            case SQL_C_SHORT:
-            case SQL_C_UTINYINT:
-            case SQL_C_USHORT:
-            case SQL_C_SBIGINT:
-            case SQL_C_SLONG:
-            case SQL_C_LONG:
-            case SQL_C_UBIGINT:
-            case SQL_C_ULONG:
-            case SQL_C_FLOAT:
-            case SQL_C_DOUBLE:
-                break;
-            default:
-                return false;
-        }
-    }
-    return true;
-}
+// The single-buffer AllocateParamBuffer template, its reuse overload, the
+// ParameterBinding / ExecuteBindingCache / ExecuteParamBuffers types,
+// UpdateParamBuffer, SameParameterShape, and CanCacheParameters live in
+// param_bind_cache.hpp. AllocateParamBufferArray below serves the executemany
+// array-binding path and is unrelated to the reuse cache, so it stays here.
 
 template <typename ParamType>
 ParamType* AllocateParamBufferArray(std::vector<std::shared_ptr<void>>& paramBuffers,
